- Added a context menu for the 2D plot curves. - Updated to the last CVS version of Qwt5

+++ branches/scripting/qtiplot/src/canvaspicker.h
@@ -47,7 +47,7 @@ private:
 	
 signals:
 	void showPieDialog();
-	void showPlotDialog(long);
+	void showPlotDialog(int);
 	void viewTextDialog();
 	void viewLineDialog();
 	void viewImageDialog();
